requirements.txt
flask
flask-cors
diffusers
torch
transformers
accelerate
Pillow
